+++ gateway_server/src/main/java/com/itmo/configs/SecurityConfig.java
@@ -1,54 +1,44 @@
 package com.itmo.configs;
 
-import com.itmo.services.FirmService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.password.NoOpPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@EnableWebSecurity
-@Configuration
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Autowired
-    FirmService firmService;
-
-    @Bean
-    PasswordEncoder noEncoder(){
-        return NoOpPasswordEncoder.getInstance();
-    }
-
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth
-                .userDetailsService(firmService)
-                .passwordEncoder(noEncoder());
-    }
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
-                .authorizeRequests()
-                    .antMatchers( "/", "/register", "index.html").permitAll()
-                    .anyRequest().authenticated()
-                .and()
-                .formLogin()
-                    .loginPage("/hello.xhtml")
-                    .usernameParameter("j_idt4:username")
-                    .passwordParameter("j_idt4:password")
-                    .defaultSuccessUrl("/firms/getAll", true)
-                    .permitAll()
-                .and()
-                .logout()
-                    .invalidateHttpSession(true)
-                    .deleteCookies("JSESSIONID")
-                    .permitAll()
-                .and()
-                .csrf().disable();
-    }
-}
+//
+//@EnableWebSecurity
+//@Configuration
+//public class SecurityConfig extends WebSecurityConfigurerAdapter {
+////
+////    @Autowired
+////    FirmService firmService;
+////
+////    @Bean
+////    PasswordEncoder noEncoder(){
+////        return NoOpPasswordEncoder.getInstance();
+////    }
+////
+////    @Override
+////    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+////        auth
+////                .userDetailsService(firmService)
+////                .passwordEncoder(noEncoder());
+////    }
+//
+//    @Override
+//    protected void configure(HttpSecurity http) throws Exception {
+//        http
+//                .authorizeRequests()
+//                    .antMatchers( "/*").permitAll()
+////                    .anyRequest().authenticated()
+//                .and()
+////                .formLogin()
+////                    .loginPage("/hello.xhtml")
+////                    .usernameParameter("j_idt4:username")
+////                    .passwordParameter("j_idt4:password")
+////                    .defaultSuccessUrl("/firms/getAll", true)
+////                    .permitAll()
+////                .and()
+////                .logout()
+////                    .invalidateHttpSession(true)
+////                    .deleteCookies("JSESSIONID")
+////                    .permitAll()
+////                .and()
+//                .csrf().disable();
+//    }
+//}
